# public_html/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'API Documentation for EasyWords APP'
  description: 'Here you can find detail documentation about API endpoints for EasyWords APP. Currently API supports getting short list of users and getting latest words for display in cards'
  version: 1.0.0
servers:
  -
    url: 'http://cards.sergeyem.test:8000'
paths:
...
  /api/user:
    get:
      summary: 'User Information'
      description: 'Getting authenticated user information'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Other
...
  /api/user/password:
    put:
      summary: 'Change user password'
      description: 'This API endpoint you will use for  changing user password'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    name: 'Mr. Nelson Haag'
                    email: admin@admin.com
                    email_verified_at: '2021-08-26T05:57:00.000000Z'
                    current_team_id: null
                    profile_photo_path: null
                    created_at: '2021-08-26T05:57:00.000000Z'
                    updated_at: '2021-08-26T05:57:00.000000Z'
                    deleted_at: null
                properties:
                  data:
                    type: object
                    example:
                      id: 1
                      name: 'Mr. Nelson Haag'
                      email: admin@admin.com
                      email_verified_at: '2021-08-26T05:57:00.000000Z'
                      current_team_id: null
                      profile_photo_path: null
                      created_at: '2021-08-26T05:57:00.000000Z'
                      updated_at: '2021-08-26T05:57:00.000000Z'
                      deleted_at: null
        422:
          description: 'validation error'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The given data was invalid.'
                  errors:
                    email:
                      - 'The provided credentials are incorrect.'
                properties:
                  message:
                    type: string
                    example: 'The given data was invalid.'
                  errors:
                    type: object
                    example:
                      email:
                        - 'The provided credentials are incorrect.'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: 'Current password of user.'
                  example: '"password"'
                password:
                  type: string
                  description: 'Your new password.'
                  example: '"new_password"'
                password_confirmation:
                  type: string
                  description: 'Retype your new password.'
                  example: '"new_password"'
              required:
                - current_password
                - password
                - password_confirmation
...
  /api/short:
    get:
      summary: 'Get users list'
      description: 'Receive short list of users for dropdown options. Here we get only ID and name of user.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 12
                      name: 'Ksenia Emelyanova'
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 12
                        name: 'Ksenia Emelyanova'
                    items:
                      type: object
        401:
          description: unauthorized
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Users
  /api/me:
    get:
      summary: 'Profile Info'
      description: 'Get current authenticated user data'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    name: 'Mr. Nelson Haag'
                    email: admin@admin.com
                    email_verified_at: '2021-08-26T05:57:00.000000Z'
                    current_team_id: null
                    profile_photo_path: null
                    created_at: '2021-08-26T05:57:00.000000Z'
                    updated_at: '2021-08-26T05:57:00.000000Z'
                    deleted_at: null
                properties:
                  data:
                    type: object
                    example:
                      id: 1
                      name: 'Mr. Nelson Haag'
                      email: admin@admin.com
                      email_verified_at: '2021-08-26T05:57:00.000000Z'
                      current_team_id: null
                      profile_photo_path: null
                      created_at: '2021-08-26T05:57:00.000000Z'
                      updated_at: '2021-08-26T05:57:00.000000Z'
                      deleted_at: null
        401:
          description: unauthorized
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Users
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []